+++ tests/ByteDev.Encoding.UnitTests/Base64/Base64EncoderTests.cs
@@ -10,15 +10,52 @@ namespace ByteDev.Encoding.UnitTests.Base64
         [TestFixture]
         public class CalcBase64EncodedSize
         {
-            [TestCase(-1, 0)]
-            [TestCase(0, 0)]
-            [TestCase(10, 16)]
-            [TestCase(15, 20)]
-            [TestCase(16, 24)]
-            [TestCase(50, 68)]
-            public void WhenOriginalSizeProvided_ThenReturnExpectedBase64Size(long originalSize, long expected)
+            [TestCase(-2)]
+            [TestCase(-1)]
+            [TestCase(0)]
+            public void WhenOriginalSizeIsZeroOrLess_ThenReturnZero(long originalSize)
             {
                 var result = Base64Encoder.CalcBase64EncodedSize(originalSize);
+
+                Assert.That(result, Is.Zero);
+            }
+
+            [TestCase(1, 4)]
+            [TestCase(2, 4)]
+            [TestCase(3, 4)]
+            [TestCase(4, 8)]
+            [TestCase(5, 8)]
+            [TestCase(6, 8)]
+            [TestCase(50, 68)]
+            public void WhenOriginalSizeIsGreaterThanZero_ThenReturnBase64Size(long originalSize, long expected)
+            {
+                var result = Base64Encoder.CalcBase64EncodedSize(originalSize);
+
+                Assert.That(result, Is.EqualTo(expected));
+            }
+        }
+
+        [TestFixture]
+        public class CalcOriginalSize
+        {
+            [TestCase(null)]
+            [TestCase("")]
+            public void WhenBase64StringIsNullOrEmpty_ThenReturnZero(string base64)
+            {
+                var result = Base64Encoder.CalcOriginalSize(base64);
+
+                Assert.That(result, Is.Zero);
+            }
+
+            [TestCase("QQ==", 1)]       // A
+            [TestCase("QUI=", 2)]       // AB
+            [TestCase("QUJD", 3)]       // ABC
+            [TestCase("QUJDRA==", 4)]   // ABCD
+            [TestCase("QUJDREU=", 5)]   // ABCDE
+            [TestCase("QUJDREVG", 6)]   // ABCDEF
+            public void WhenBase64StringIsValid_ThenReturnOriginalSize(string base64, long expected)
+            {
+                var result = Base64Encoder.CalcOriginalSize(base64);
 
                 Assert.That(result, Is.EqualTo(expected));
             }